txflow/producer/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
confluent-kafka==2.4.0
psycopg2-binary==2.9.9
pydantic==2.7.0
python-dotenv==1.0.1

